+++ ppvm-python/src/ppvm/__init__.py
@@ -1,1 +1,1 @@
-from .wrapper import PauliSum as PauliSum
+from .paulisum import PauliSum as PauliSum
