+++ ppvm-python/src/ppvm/paulisum.py
